Lua method for TemplateData
The mw.ext.TemplateData.load() method has been added to Lua
to get data for the specified or current page.

Extending the previous changeset 7 with additional assertions of args,
faster casting of objects, localization of error messages, additional
cache of previous lookups, and some cleanups.

Cache now handles multiple values. This is not necessary for this
changeset, but is necessary for general use when multiple return
values are allowed as in Lua.

No configuration for linting with luacheck.

Fixed name in language files. (There are more entries that should
be fixed.)

Fixed a bogus if-clause.

Fixed passing of config to the create()

Fixed a direct access to a member.

Fixed an erronuous recount.

Added tracking of page changes.

Added method for proxying the cached value.

Bug: T107119
Change-Id: I1667c15758947218ae69d856194002d7febda2b9